Google Analytics: migrate the module code into a package.
